Recursively create the node['openvpn']['key_dir']
Useful when wanting to put the Key on a seperate partition (e.g. in AWS where the root volume isn't encrypted but another mount point is).

Without `recursive true` it is necessary to create parent directories to the `key_dir` path seperately, so it seems reasonable to recursively create them (equiv of `mkdir -p`)